refactor(Player): Money에 대한 역할을 Play에 할당
