+++ src/main/java/blackjack/domain/participant/Participant.java
@@ -1,6 +1,5 @@
 package blackjack.domain.participant;
 
-import blackjack.domain.Money;
 import blackjack.domain.Name;
 import blackjack.domain.card.Cards;
 import blackjack.domain.card.Deck;
@@ -14,18 +13,12 @@ public abstract class Participant {
 
     private final Cards cards;
     private final Name name;
-    private final Money money;
     private Status status;
 
     public Participant(String name) {
         this.cards = new Cards();
         this.name = new Name(name);
         this.status = Status.HIT;
-        this.money = new Money();
-    }
-
-    public void initMoney(final int bettingAmount) {
-        money.init(bettingAmount);
     }
 
     public void initCards(Deck deck) {
@@ -64,21 +57,5 @@ public abstract class Participant {
 
     public boolean isBlackjack() {
         return cards.sumValue() == BLACKJACK_STANDARD && cards.getValue().size() == 2;
-    }
-
-    public void updateBlackjackPrize() {
-        money.multiply();
-    }
-
-    public void updatePushPrize() {
-        money.toZero();
-    }
-
-    public void updateNegativePrize() {
-        money.toNegative();
-    }
-
-    public int getPrize() {
-        return money.getValue();
     }
 }
